+++ examples/async/get_status_auto_update.py
@@ -15,13 +15,13 @@ async def async_callback_example(component):
 async def main():
     async with Device("pi.local", 8080) as device:
         print("Starting auto-update")
-        await device.start_auto_update(
+        await device.auto_update_start(
             update_callback=print_component,
             update_callback_async=async_callback_example,
         )
         await asyncio.sleep(20.0)
         print("Stopping auto-update")
-        await device.stop_auto_update()
+        await device.auto_update_stop()
 
 
 if __name__ == "__main__":