+++ CHANGES.rst
@@ -8,6 +8,7 @@ v0.0.5
 - Rename ``pupil_labs.realtime_api.basic.Control`` to :py:class:`pupil_labs.realtime_api.basic.Device`.
 - Rename ``pupil_labs.realtime_api.control.Control`` to :py:class:`pupil_labs.realtime_api.device.Device`.
 - Rename ``pupil_labs.realtime_api.models.DiscoveredDevice`` to :py:class:`pupil_labs.realtime_api.models.DiscoveredDeviceInfo`.
+- Add sensor property accessors to :py:class:`pupil_labs.realtime_api.basic.Device`.
 
 v0.0.4
 ######